chore: fix file duplication
